Add pom.xml to election-api-java project
When running assessments on BBC laptops it is straight forward to
have IDEs setup and configured to work with gradle projects, however
we had a candidate who was unfamiliar with gradle and unable to
integrate it with their prefered IDE (Netbeans).

This commit will add a pom.xml file that can be used by maven to manage
the project as an alternative to the typical gradle build we expect
candidates to use, at the cost of an additional file to maintain when
we update dependacies.