Add maven wrapper and usage instructions
To add maven as an alternative build tool to gradle we need
to provide candidates usage instructions and support equal to
that of gradle.

Rather than get too much into installing and configuring maven
for different platforms, this commit adds a mvn wrapper much
like the gradle wrapper which can be used to test, build and
run the coding exercise.