Replace JsonProperty annotation with POJO setters in Java API model
In a previous commit, the Java API assessment was modified to remove
reference to the lombok annotations that provided builder/property
functionality. In the same change, the JsonProperty annotation was
introduced to provide Serialisation/Deserialisation to JSON. To further
reduce the overhead and prior knowledge required by interview
candidates, it would be an improvement to remove the need for the
JsonProperty annotation and rely on plain Java objects where possible.

This commit removes the JsonProperty and replaces the functionality by
providing a default constructor and setters for the properties to be
deserialised. Spring boot will automatically serialise objects in the
controller (using fasterxml Jackson) as the class is annotated with
RestController. A trade-off of removing the annotation is that we have
lost immutability; however the code may be more accessible to
candidates using this approach.